refactor: update tests
